renamed the package to nysure
