+++ src/main/java/se/kth/sda6/nysure/auth/AuthRequest.java
@@ -1,4 +1,4 @@
-package se.kth.sda6.skeleton.auth;
+package se.kth.sda6.nysure.auth;
 
 public class AuthRequest {
     private String email;
