+++ src/main/java/se/kth/sda6/nysure/user/UserRepository.java
@@ -1,4 +1,4 @@
-package se.kth.sda6.skeleton.user;
+package se.kth.sda6.nysure.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
