+++ src/main/resources/application.properties
@@ -1,8 +1,8 @@
 spring.jpa.database=POSTGRESQL
 spring.datasource.platform=postgres
-spring.datasource.url=jdbc:postgresql://localhost:5433/skeleton
-spring.datasource.username=skeleton_user
-spring.datasource.password=skeleton_pass
+spring.datasource.url=jdbc:postgresql://localhost:5433/nysure
+spring.datasource.username=nysure_user
+spring.datasource.password=nysure_pass
 spring.jpa.show-sql=true
 spring.jpa.generate-ddl=true
 spring.jpa.hibernate.ddl-auto=create
