+++ src/test/java/se/kth/sda6/skeleton/NysureApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class SkeletonApplicationTests {
+class NysureApplicationTests {
 
 	@Test
 	void contextLoads() {
